move SGE-related setup to new SGE plugin
updated Cluster to automatically run new SGE plugin if disable_queue is
false. this maintains the default behaviour for StarCluster (ie creating
SGE-enabled clusters)
