+++ starcluster/cluster.py
@@ -18,6 +18,7 @@ from starcluster import exception
 from starcluster import progressbar
 from starcluster import clustersetup
 from starcluster.node import Node
+from starcluster.plugins import sge
 from starcluster.utils import print_timing
 from starcluster.templates import user_msgs
 from starcluster.logger import log
@@ -816,14 +817,18 @@ class Cluster(object):
             self.create_nodes(aliases, count=len(aliases))
         self.wait_for_cluster(msg="Waiting for node(s) to come up...")
         log.debug("Adding node(s): %s" % aliases)
-        default_plugin = clustersetup.DefaultClusterSetup(self.disable_queue,
-                                                          self.disable_threads)
+        default_plugin = clustersetup.DefaultClusterSetup(self.disable_threads)
+        if not self.disable_queue:
+            sge_plugin = sge.SGEPlugin(self.disable_threads)
         for alias in aliases:
             node = self.get_node_by_alias(alias)
-            default_plugin.on_add_node(
-                node, self.nodes, self.master_node,
-                self.cluster_user, self.cluster_shell,
-                self.volumes)
+            default_plugin.on_add_node(node, self.nodes, self.master_node,
+                                       self.cluster_user, self.cluster_shell,
+                                       self.volumes)
+            if not self.disable_queue:
+                sge_plugin.on_add_node(node, self.nodes, self.master_node,
+                                       self.cluster_user, self.cluster_shell,
+                                       self.volumes)
             self.run_plugins(method_name="on_add_node", node=node)
 
     def remove_node(self, node, terminate=True):
@@ -836,17 +841,21 @@ class Cluster(object):
         """
         Remove a list of nodes from this cluster
         """
-        default_plugin = clustersetup.DefaultClusterSetup(self.disable_queue,
-                                                          self.disable_threads)
+        default_plugin = clustersetup.DefaultClusterSetup(self.disable_threads)
+        if not self.disable_queue:
+            sge_plugin = sge.SGEPlugin(self.disable_threads)
         for node in nodes:
             if node.is_master():
                 raise exception.InvalidOperation("cannot remove master node")
             self.run_plugins(method_name="on_remove_node",
                              node=node, reverse=True)
-            default_plugin.on_remove_node(
-                node, self.nodes, self.master_node,
-                self.cluster_user, self.cluster_shell,
-                self.volumes)
+            if not self.disable_queue:
+                sge_plugin.on_remove_node(node, self.nodes, self.master_node,
+                                          self.cluster_user,
+                                          self.cluster_shell, self.volumes)
+            default_plugin.on_remove_node(node, self.nodes, self.master_node,
+                                          self.cluster_user,
+                                          self.cluster_shell, self.volumes)
             if not terminate:
                 continue
             if node.spot_id:
@@ -1431,9 +1440,12 @@ class Cluster(object):
         log.info("Setting up the cluster...")
         if self.volumes:
             self.attach_volumes_to_master()
-        default_plugin = clustersetup.DefaultClusterSetup(self.disable_queue,
-                                                          self.disable_threads)
+        default_plugin = clustersetup.DefaultClusterSetup(self.disable_threads)
         default_plugin.run(self.nodes, self.master_node, self.cluster_user,
+                           self.cluster_shell, self.volumes)
+        if not self.disable_queue:
+            sge_plugin = sge.SGEPlugin(self.disable_threads)
+            sge_plugin.run(self.nodes, self.master_node, self.cluster_user,
                            self.cluster_shell, self.volumes)
         self.run_plugins()
