+++ starcluster/clustersetup.py
@@ -3,12 +3,10 @@
 """
 clustersetup.py
 """
-import re
 import posixpath
 
 from starcluster import threadpool
 from starcluster.utils import print_timing
-from starcluster.templates import sge
 from starcluster.logger import log
 
 
@@ -61,14 +59,12 @@ class DefaultClusterSetup(ClusterSetup):
     """
     Default ClusterSetup implementation for StarCluster
     """
-    def __init__(self, disable_queue=False, disable_threads=False,
-                 num_threads=20):
+    def __init__(self, disable_threads=False, num_threads=20):
         self._nodes = None
         self._master = None
         self._user = None
         self._user_shell = None
         self._volumes = None
-        self._disable_queue = disable_queue
         self._disable_threads = disable_threads
         self._num_threads = num_threads
         self._pool = None
@@ -283,8 +279,6 @@ class DefaultClusterSetup(ClusterSetup):
 
     def _get_nfs_export_paths(self):
         export_paths = ['/home']
-        if not self._disable_queue:
-            export_paths.append('/opt/sge6')
         for vol in self._volumes:
             vol = self._volumes[vol]
             mount_path = vol.get('mount_path')
@@ -292,10 +286,10 @@ class DefaultClusterSetup(ClusterSetup):
                 export_paths.append(mount_path)
         return export_paths
 
-    def _mount_nfs_shares(self, nodes):
+    def _mount_nfs_shares(self, nodes, export_paths=None):
         # setup /etc/fstab and mount each nfs share on each node
         master = self._master
-        export_paths = self._get_nfs_export_paths()
+        export_paths = export_paths or self._get_nfs_export_paths()
         for node in nodes:
             mount_map = node.get_mount_map()
             mount_paths = []
@@ -303,8 +297,8 @@ class DefaultClusterSetup(ClusterSetup):
                 network_device = "%s:%s" % (master.alias, path)
                 if network_device in mount_map:
                     mount_path, type, options = mount_map.get(network_device)
-                    log.debug(('nfs share %s already mounted to %s on ' + \
-                               'node %s, skipping...') % \
+                    log.debug('nfs share %s already mounted to %s on '
+                              'node %s, skipping...' %
                               (network_device, mount_path, node.alias))
                 else:
                     mount_paths.append(path)
@@ -313,100 +307,19 @@ class DefaultClusterSetup(ClusterSetup):
         self.pool.wait(numtasks=len(nodes))
 
     @print_timing("Setting up NFS")
-    def _setup_nfs(self, nodes=None, start_server=True):
+    def _setup_nfs(self, nodes=None, start_server=True, export_paths=None):
         """
-        Share /home, /opt/sge6, and all EBS mount paths via NFS to all nodes
+        Share /home and all EBS mount paths via NFS to all nodes
         """
         log.info("Configuring NFS...")
         master = self._master
-        if not self._disable_queue and not master.ssh.isdir('/opt/sge6'):
-            # copy fresh sge installation files to /opt/sge6
-            master.ssh.execute('cp -r /opt/sge6-fresh /opt/sge6')
-            master.ssh.execute('chown -R %(user)s:%(user)s /opt/sge6' % \
-                               {'user': self._user})
         # setup /etc/exports and start nfsd on master node
         nodes = nodes or self.nodes
-        export_paths = self._get_nfs_export_paths()
+        export_paths = export_paths or self._get_nfs_export_paths()
         if start_server:
             master.start_nfs_server()
         master.export_fs_to_nodes(nodes, export_paths)
-        self._mount_nfs_shares(nodes)
-
-    def _create_sge_pe(self, name="orte", nodes=None, queue="all.q"):
-        """
-        Create or update an SGE parallel environment
-
-        name - name of parallel environment
-        nodes - list of nodes to include in the parallel environment
-                (default: all)
-        queue - configure queue to use the new parallel environment
-        """
-        mssh = self._master.ssh
-        pe_exists = mssh.get_status('qconf -sp %s' % name, source_profile=True)
-        pe_exists = pe_exists == 0
-        if not pe_exists:
-            log.info("Creating SGE parallel environment '%s'" % name)
-        else:
-            log.info("Updating SGE parallel environment '%s'" % name)
-        # iterate through each machine and count the number of processors
-        nodes = nodes or self._nodes
-        num_processors = sum(self.pool.map(lambda n: n.num_processors, nodes))
-        penv = mssh.remote_file("/tmp/pe.txt")
-        print >> penv, sge.sge_pe_template % (name, num_processors)
-        penv.close()
-        if not pe_exists:
-            mssh.execute("qconf -Ap %s" % penv.name, source_profile=True)
-        else:
-            mssh.execute("qconf -Mp %s" % penv.name, source_profile=True)
-        if queue:
-            log.info("Adding parallel environment '%s' to queue '%s'" %
-                     (name, queue))
-            mssh.execute('qconf -mattr queue pe_list "%s" %s' % (name, queue),
-                         source_profile=True)
-
-    def _setup_sge(self):
-        """
-        Install Sun Grid Engine with a default parallel
-        environment on StarCluster
-        """
-        # generate /etc/profile.d/sge.sh for each node
-        for node in self._nodes:
-            conn = node.ssh
-            conn.execute('pkill -9 sge', ignore_exit_status=True)
-            conn.execute('rm /etc/init.d/sge*', ignore_exit_status=True)
-            sge_profile = conn.remote_file("/etc/profile.d/sge.sh")
-            arch = conn.execute("/opt/sge6/util/arch")[0]
-            print >> sge_profile, sge.sgeprofile_template % {'arch': arch}
-            sge_profile.close()
-        # setup sge auto install file
-        master = self._master
-        default_cell = '/opt/sge6/default'
-        if master.ssh.isdir(default_cell):
-            log.info("Removing previous SGE installation...")
-            master.ssh.execute('rm -rf %s' % default_cell)
-            master.ssh.execute('exportfs -fr')
-        mconn = master.ssh
-        admin_list = ' '.join(map(lambda n: n.alias, self._nodes))
-        exec_list = admin_list
-        submit_list = admin_list
-        ec2_sge_conf = mconn.remote_file("/opt/sge6/ec2_sge.conf")
-        # TODO: add sge section to config values for some of the below
-        conf = sge.sgeinstall_template % (admin_list, exec_list, submit_list)
-        print >> ec2_sge_conf, conf
-        ec2_sge_conf.close()
-        # installs sge in /opt/sge6 and starts qmaster/schedd on master node
-        log.info("Installing Sun Grid Engine...")
-        mconn.execute('cd /opt/sge6 && TERM=rxvt ./inst_sge -m -x -noremote '
-                      '-auto ./ec2_sge.conf', silent=True, only_printable=True)
-        # set all.q shell to bash
-        mconn.execute('qconf -mattr queue shell "/bin/bash" all.q',
-                      source_profile=True)
-        for node in self.nodes:
-            self._add_sge_administrative_host(node)
-            self._add_sge_submit_host(node)
-            self.pool.simple_job(self._add_to_sge, (node,), jobid=node.alias)
-        self.pool.wait(numtasks=len(self.nodes))
-        self._create_sge_pe()
+        self._mount_nfs_shares(nodes, export_paths=export_paths)
 
     def run(self, nodes, master, user, user_shell, volumes):
         """Start cluster configuration"""
@@ -423,8 +336,6 @@ class DefaultClusterSetup(ClusterSetup):
             self._setup_etc_hosts()
             self._setup_nfs()
             self._setup_passwordless_ssh()
-            if not self._disable_queue:
-                self._setup_sge()
         finally:
             self.pool.shutdown()
 
@@ -442,53 +353,6 @@ class DefaultClusterSetup(ClusterSetup):
             n.remove_from_known_hosts('root', [node])
             n.remove_from_known_hosts(self._user, [node])
 
-    def _remove_from_sge(self, node):
-        master = self._master
-        master.ssh.execute('qconf -shgrp @allhosts > /tmp/allhosts',
-                           source_profile=True)
-        hgrp_file = master.ssh.remote_file('/tmp/allhosts', 'r')
-        contents = hgrp_file.read().splitlines()
-        hgrp_file.close()
-        c = []
-        for line in contents:
-            line = line.replace(node.alias, '')
-            c.append(line)
-        hgrp_file = master.ssh.remote_file('/tmp/allhosts_new', 'w')
-        hgrp_file.writelines('\n'.join(c))
-        hgrp_file.close()
-        master.ssh.execute('qconf -Mhgrp /tmp/allhosts_new',
-                           source_profile=True)
-        master.ssh.execute('qconf -sq all.q > /tmp/allq', source_profile=True)
-        allq_file = master.ssh.remote_file('/tmp/allq', 'r')
-        contents = allq_file.read()
-        allq_file.close()
-        c = [l.strip() for l in contents.splitlines()]
-        s = []
-        allq = []
-        for l in c:
-            if l.startswith('slots') or l.startswith('['):
-                s.append(l)
-            else:
-                allq.append(l)
-        regex = re.compile(r"\[%s=\d+\],?" % node.alias)
-        slots = []
-        for line in s:
-            line = line.replace('\\', '')
-            slots.append(regex.sub('', line))
-        allq.append(''.join(slots))
-        f = master.ssh.remote_file('/tmp/allq_new', 'w')
-        allq[-1] = allq[-1].strip()
-        if allq[-1].endswith(','):
-            allq[-1] = allq[-1][:-1]
-        f.write('\n'.join(allq))
-        f.close()
-        master.ssh.execute('qconf -Mq /tmp/allq_new', source_profile=True)
-        master.ssh.execute('qconf -dconf %s' % node.alias, source_profile=True)
-        master.ssh.execute('qconf -de %s' % node.alias, source_profile=True)
-        node.ssh.execute('pkill -9 sge_execd')
-        nodes = filter(lambda n: n.alias != node.alias, self._nodes)
-        self._create_sge_pe(nodes=nodes)
-
     def on_remove_node(self, node, nodes, master, user, user_shell, volumes):
         self._nodes = nodes
         self._master = master
@@ -496,9 +360,6 @@ class DefaultClusterSetup(ClusterSetup):
         self._user_shell = user_shell
         self._volumes = volumes
         log.info("Removing node %s (%s)..." % (node.alias, node.id))
-        if not self._disable_queue:
-            log.info("Removing %s from SGE" % node.alias)
-            self._remove_from_sge(node)
         log.info("Removing %s from known_hosts files" % node.alias)
         self._remove_from_known_hosts(node)
         log.info("Removing %s from /etc/hosts" % node.alias)
@@ -510,23 +371,6 @@ class DefaultClusterSetup(ClusterSetup):
         user = self._master.getpwnam(self._user)
         uid, gid = user.pw_uid, user.pw_gid
         self._add_user_to_nodes(uid, gid, nodes=[node])
-
-    def _add_sge_submit_host(self, node):
-        mssh = self._master.ssh
-        mssh.execute('qconf -as %s' % node.alias, source_profile=True)
-
-    def _add_sge_administrative_host(self, node):
-        mssh = self._master.ssh
-        mssh.execute('qconf -ah %s' % node.alias, source_profile=True)
-
-    def _add_to_sge(self, node):
-        # generate /etc/profile.d/sge.sh
-        sge_profile = node.ssh.remote_file("/etc/profile.d/sge.sh")
-        arch = node.ssh.execute("/opt/sge6/util/arch")[0]
-        print >> sge_profile, sge.sgeprofile_template % {'arch': arch}
-        sge_profile.close()
-        node.ssh.execute('cd /opt/sge6 && TERM=rxvt ./inst_sge -x -noremote '
-                         '-auto ./ec2_sge.conf')
 
     def on_add_node(self, node, nodes, master, user, user_shell, volumes):
         self._nodes = nodes
@@ -540,9 +384,3 @@ class DefaultClusterSetup(ClusterSetup):
         self._create_user(node)
         self._setup_scratch(nodes=[node])
         self._setup_passwordless_ssh(nodes=[node])
-        if not self._disable_queue:
-            log.info("Adding %s to SGE" % node.alias)
-            self._add_sge_administrative_host(node)
-            self._add_sge_submit_host(node)
-            self._add_to_sge(node)
-            self._create_sge_pe()
